start apps/expo router shaz from scratch
