+++ apps/expo/app-prev/(tabs)/index.tsx
