+++ apps/expo/app-prev/_layout.tsx
@@ -1,0 +1,54 @@
+import {
+  SourceCodePro_400Regular,
+  useFonts,
+} from '@expo-google-fonts/source-code-pro'
+import FontAwesome from '@expo/vector-icons/FontAwesome'
+import { SplashScreen, Stack } from 'expo-router'
+import { StatusBar } from 'expo-status-bar'
+import { useEffect } from 'react'
+import ErrorToastContainer from '@bacons/expo-metro-runtime/error-overlay'
+import { Provider } from 'app/provider'
+
+export { ErrorBoundary } from 'expo-router'
+
+export const unstable_settings = {
+  initialRouteName: '(tabs)',
+}
+
+export default function RootLayout() {
+  const [loaded, error] = useFonts({
+    Inter: require('@tamagui/font-inter/otf/Inter-Medium.otf'),
+    InterBold: require('@tamagui/font-inter/otf/Inter-Bold.otf'),
+    ...FontAwesome.font,
+    SourceCodePro_400Regular,
+  })
+
+  // Expo Router uses Error Boundaries to catch errors in the navigation tree.
+  useEffect(() => {
+    if (error) throw error
+  }, [error])
+
+  return (
+    <>
+      {!loaded && <SplashScreen />}
+      {loaded && <RootLayoutNav />}
+    </>
+  )
+}
+
+function RootLayoutNav() {
+  return (
+    <>
+      <ErrorToastContainer>
+        <Provider>
+          <Stack>
+            <Stack.Screen name="index" options={{ headerShown: false }} />
+            <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
+            <Stack.Screen name="modal" options={{ presentation: 'modal' }} />
+          </Stack>
+        </Provider>
+      </ErrorToastContainer>
+      <StatusBar style="light" />
+    </>
+  )
+}
