+++ apps/expo/app-prev/index.tsx
@@ -1,0 +1,11 @@
+import { HomeScreen } from 'app/features/home/HomeScreen'
+import { StatusBar } from 'expo-status-bar'
+
+export default function Home() {
+  return (
+    <>
+      <HomeScreen />
+      <StatusBar style="light" />
+    </>
+  )
+}
