+++ apps/expo/app/index.tsx
@@ -1,11 +1,5 @@
 import { HomeScreen } from 'app/features/home/HomeScreen'
-import { StatusBar } from 'expo-status-bar'
 
-export default function Home() {
-  return (
-    <>
-      <HomeScreen />
-      <StatusBar style="light" />
-    </>
-  )
+export default function () {
+  return <HomeScreen />
 }
